+++ src/routes/notebooks/[id]/+page.svelte
@@ -216,6 +216,9 @@
 			contents={data.contents}
 			allPeople={data.allPeople}
 			categories={data.categories}
+			inventoryCategories={data.inventoryCategories}
+			workoutCategories={data.workoutCategories}
+			parsers={data.parsers}
 			currency={data.currency}
 			pickableNotebooks={data.pickableNotebooks}
 			areas={data.areas}